# Fix/symbolic invariant vectors