--- a/src/Characters/echelon_form.jl
+++ b/src/Characters/echelon_form.jl
@@ -1,44 +1,58 @@
-# Version over exact field:
-function row_echelon_form!(A::AbstractMatrix{T}) where {T<:Union{FiniteFields.GF, Cyclotomic{<:Rational}, Rational}}
-    pivots = Int[]
-    pos = 0
-    for i = 1:size(A, 2)
-        j = findnext(x -> !iszero(x), @view(A[:, i]), pos+1)
-        j === nothing && continue
-        pos += 1
-        push!(pivots, i)
+Base.@propagate_inbounds function _swap_rows!(A::AbstractMatrix, i, j)
+    @boundscheck @assert 1 ≤ i ≤ size(A, 1)
+    @boundscheck @assert 1 ≤ j ≤ size(A, 1)
 
-        # swap the rows so that
-        if (pos != j)
-            for k in 1:size(A, 2)
-                A[pos, k], A[j, k] = A[j, k], A[pos, k]
-            end
-        end
-        # A[pos, :] is the row with leading nz in i-th column
+    i == j && return A
 
-        w = inv(A[pos, i])
-        # all columns to the left of i are zeroed (below pivots) so we start at i
-        for colidx in i:size(A, 2)
-            A[pos, colidx] *= w
-        end
-        # A[pos, i] is 1 now
-
-        # zero the whole i-th column above and below pivot:
-        for rowidx = 1:size(A, 1)
-            rowidx == pos && continue # don't zero the active row
-            v = A[rowidx, i]
-            iszero(v) && continue
-            for k in i:size(A, 2) # to the left of pivot everything is zero
-                A[rowidx, k] -= v * A[pos, k]
-            end
-        end
+    @inbounds for col_idx in 1:size(A, 2)
+        A[i, col_idx], A[j, col_idx] = A[j, col_idx], A[i, col_idx]
     end
-    return A, pivots
+    return A
 end
 
-function _findmax(f, A::AbstractArray)
+Base.@propagate_inbounds function _mul_row!(A::AbstractMatrix, val, row_idx, starting_at = 1)
+    @boundscheck @assert 1 ≤ starting_at ≤ size(A, 2)
+
+    @inbounds for col_idx in starting_at:size(A, 2)
+        A[row_idx, col_idx] *= val
+    end
+    return A
+end
+
+# Version over exact field:
+
+Base.@propagate_inbounds function _find_pivot(A::AbstractMatrix, col_idx, starting_at = 1)
+    k = findnext(!iszero, @view(A[:, col_idx]), starting_at)
+    isnothing(k) && return false, col_idx
+    return true, k
+end
+
+Base.@propagate_inbounds function _reduce_column_by_pivot!(
+    A::AbstractMatrix,
+    row_idx,
+    col_idx,
+    starting_at = 1,
+)
+    @boundscheck checkbounds(A, row_idx, col_idx)
+    @boundscheck 1 ≤ starting_at ≤ size(A, 2)
+
+    Threads.@threads for ridx in 1:size(A, 1)
+        ridx == row_idx && continue
+        v = A[ridx, col_idx]
+        iszero(v) && continue
+        @inbounds for cidx in starting_at:size(A, 2)
+            A[ridx, cidx] -= v * A[row_idx, cidx]
+        end
+    end
+    return A
+end
+
+# version over AbstractFloat
+
+Base.@propagate_inbounds function __findmax(f, A::AbstractArray)
+    @assert !isempty(A)
     mval, midx = f(first(A)), firstindex(A)
-    for idx in eachindex(A)
+    @inbounds for idx in eachindex(A)
         if (v = f(A[idx])) > mval
             mval, midx = v, idx
         end
@@ -46,50 +60,74 @@ function _findmax(f, A::AbstractArray)
     return mval, midx
 end
 
-function row_echelon_form!(A::AbstractMatrix{T}, atol=eps(T)*max(size(A)...)) where {T<:AbstractFloat}
-    pivots = Int[]
-    pos = 0
-    sgn = 1.0
-    for i = 1:size(A, 2)
-        # find the largest entry in the column below the last pivot
-        mval, midx = _findmax(abs, @view(A[pos+1:end, i]))
-        # j = findnext(x -> !iszero(x), , pos+1)
-        if mval < atol # the largest entry is below threshold so we zero everything in the column
-            A[pos+1:end, i] .= zero(T)
+Base.@propagate_inbounds function _find_pivot(
+    A::AbstractMatrix{<:AbstractFloat},
+    col_idx,
+    starting_at = 1;
+    atol = eps(eltype(A))*max(size(A)...),
+)
+    isempty(starting_at:size(A, 1)) && return false, starting_at
+    @boundscheck @assert 1 ≤ starting_at
+    # find the largest entry in the column below the last pivot
+    @boundscheck @assert 1 ≤ col_idx ≤ size(A, 2)
+
+    mval, midx = __findmax(abs, @view(A[starting_at:end, col_idx]))
+    if mval < atol # the largest entry is below threshold so we zero everything in the column
+        @inbounds for ridx in starting_at:size(A, 1)
+            A[ridx, col_idx] = zero(eltype(A))
+        end
+        return false, starting_at
+    end
+    return true, oftype(starting_at, starting_at + midx - 1)
+end
+
+Base.@propagate_inbounds function _reduce_column_by_pivot!(
+    A::AbstractMatrix{T},
+    row_idx,
+    col_idx,
+    starting_at = 1,
+) where {T<:AbstractFloat}
+
+    @boundscheck checkbounds(A, row_idx, col_idx)
+    @boundscheck 1 ≤ starting_at ≤ size(A, 2)
+
+    Threads.@threads for ridx in 1:size(A, 1)
+        ridx == row_idx && continue
+        @inbounds v = A[ridx, col_idx]
+        if abs(v) < eps(T)
+            @inbounds A[ridx, col_idx] = zero(T)
             continue
         end
-        j = midx+pos
-        pos += 1
-        push!(pivots, i)
-
-        # swap the rows so that
-        if (pos != j)
-            for k in 1:size(A, 2)
-                A[pos, k], A[j, k] = A[j, k], A[pos, k]
-            end
-        end
-        # A[pos, :] is the row with leading nz in i-th column
-
-        w = inv(A[pos, i])
-        sgn *= sign(w)
-        # all columns to the left of i are zeroed (below pivots) so we start at i
-        for colidx in i:size(A, 2)
-            A[pos, colidx] *= w
-        end
-        # A[pos, i] is 1 now
-
-        # zero the whole i-th column above and below pivot:
-        for rowidx = 1:size(A, 1)
-            rowidx == pos && continue # don't zero the active row
-            v = A[rowidx, i]
-            iszero(v) && continue
-            for k in i:size(A, 2) # to the left of pivot everything is zero
-                A[rowidx, k] -= v * A[pos, k]
-            end
+        for cidx in starting_at:size(A, 2)
+            @inbounds A[ridx, cidx] -=
+                cidx == col_idx ? A[ridx, cidx] : v * A[row_idx, cidx]
         end
     end
-    if sgn < 0
-        A .*= sgn
+    return A
+end
+
+Base.@propagate_inbounds function row_echelon_form!(A::AbstractMatrix)
+    pivots = Int[]
+    row_idx = 0 # also: current_rank
+    @inbounds for col_idx in 1:size(A, 2)
+        found, j = _find_pivot(A, col_idx, row_idx + 1)
+        found || continue
+        row_idx += 1
+        push!(pivots, col_idx)
+
+        # swap the rows so that  A[row_idx, :] is the row with leading nz in
+        # i-th column
+        _swap_rows!(A, row_idx, j)
+
+        w = inv(A[row_idx, col_idx])
+
+        # multiply A[row_idx, :] by w
+        _mul_row!(A, w, row_idx, col_idx)
+        # A[current_rank, col_idx] is 1 now
+
+        # zero the whole col_idx-th column above and below pivot:
+        # to the left of col_idx everything is zero
+        _reduce_column_by_pivot!(A, row_idx, col_idx, col_idx)
     end
     return A, pivots
 end