… .github/workflows/ci.yml
… Project.toml
… examples/Manifest.toml
… examples/action_polynomials.jl
… examples/ex_C2_linear.jl
… examples/ex_S4.jl
… examples/ex_motzkin.jl
… examples/ex_robinson_form.jl
… examples/run_examples.jl
… src/minimal_projections.jl
… src/sa_basis.jl
… src/wedderburn_decomposition.jl
… test/action_dihedral.jl
… test/action_linear.jl
… test/action_permutation.jl
… test/characters.jl
… test/runtests.jl
… test/sa_basis.jl