this gets rid of numerical linear algebra computations for invariant vectors using the technology we've already had: projections in the group ring.

`invariant_vectors` constitute basis for the space of `G`-invariant vectors `⟨basis_full⟩^G`, hence their computation is the same as finding `image_basis` for the projection defined by the trivial character. If action is by permutation orbits are disjoint, hence orthogonal. Otherwise we call single `svd` at the end to orthogonalize the whole basis. Bonus: no zero vectors sneak in ;)  

a baby example with `length(basis_full)=370` and `order(G) = 48` gives:
```julia
# before
@btime invariant_vectors(H, act, basis_full);
117.466 ms (594156 allocations: 120.26 MiB)
# after
tbl = CharacterTable(Rational{Int}, H) # we compute tbl in WedderburnDecomposition anyway
@btime invariant_vectors(tbl, act, basis_full);
 33.763 ms (537076 allocations: 30.36 MiB)
```

CC: @thinh-le