@thinh-le 2021-12-14
@kalmarek may need to materialize eachrow (see my repo for full error message in an example)?

`trivial_character(chtbl::CharacterTable) =
	   Character(chtbl, findfirst(r->all(isone, r), collect(eachrow(chtbl))))`