@kalmarek 2021-12-14
@blegat locally I can't even do `using SymbolicWedderburn` on `julia-1.0`. any tips on how to approach this? Or should I just drop compatibility with `1.0` and bump to `0.3.0`?