@kalmarek 2021-12-22
Breaking changes:
* `ExtensionHomomorphism`s no longer contain `features/rfeatures`. These just established a bijection and a similar functionality is provided by `StarAlgebras.Basis` so I swapped those for a single `basis`. All other functionality is left intact.


@blegat