@kalmarek 2022-01-03
Much work went into this without much of gain since `f4a60ef`...

The only thing that went well is that despite code complication I didn't break the performance ;)

* before
```julia
julia> include("bench_actions.jl")
  Activating project at `~/.julia/dev/SymbolicWedderburn`
┌ Info: invariant_vectors: perm action
│   length(SymGbas) = 1364
│   action = OnLetters()
│   SymG =
│    Permutation group on 2 generators of order 24 generated by
│     (1,2,3,4)
└     (1,2)
  127.996 ms (34889 allocations: 34.21 MiB)

┌ Info: invariant_vectors: linear action:
│   length(DihGbas) = 5151
│   daction = DihedralAction()
└   DihG = DihedralGroup(4)
  473.784 ms (5393671 allocations: 315.35 MiB)
[ Info: done!

```
* after
```julia
julia> include("bench_actions.jl")
  Activating project at `~/.julia/dev/SymbolicWedderburn`
┌ Info: invariant_vectors: perm action
│   length(SymGbas) = 1364
│   action = OnLetters()
│   SymG =
│    Permutation group on 2 generators of order 24 generated by
│     (1,2,3,4)
└     (1,2)
  127.266 ms (34728 allocations: 4.62 MiB)
┌ Info: invariant_vectors: linear action:
│   length(DihGbas) = 5151
│   daction = DihedralAction()
└   DihG = DihedralGroup(4)
  456.136 ms (5371646 allocations: 303.28 MiB)
```

but things should be much stabler now.

btw. of course a much faster version for "ByPermutation" is available:
```julia
┌ Info: invariant_vectors: perm action
│   length(SymGbas) = 1364
│   action = OnLetters()
│   SymG =
│    Permutation group on 2 generators of order 24 generated by
│     (1,2,3,4)
└     (1,2)
  856.324 μs (5765 allocations: 596.38 KiB)
```
@thinh-le  I guess your chance at tackling large problems is to come up with your own
```julia
function invariant_vectors(
    tbl::Characters.CharacterTable,
    act::YourAction,
    basis::StarAlgebras.Basis{T,I},
)
    ....
end
```
nothing generic and efficient can be done here ;)
(or prove me wrong ;D)