@kalmarek 2022-01-03
the failures on MacOS seem to be related to numerical instability; I will investigate if they occur again